Prepare for CPML optimisation :
we want to renumber PML elements so that all PML/X are contiguous, then
PML/Y, PML/Z and the last batch containing PML in several directions.

We want this to ensure that a vector always contain the same type of single
direction PML elements (eg 4x PML in X), of course we should allow holes
in between elements of different 'types'. The cases of PML in several directions
will not be optimized.